+++ src/components/ConfigEditor.tsx
@@ -1,5 +1,5 @@
-import React, { ChangeEvent } from 'react';
-import { InlineField, SecretInput, Input, Label } from '@grafana/ui';
+import React, { ChangeEvent, useState } from 'react';
+import { InlineField, SecretInput, Input, Label, Alert } from '@grafana/ui';
 import { DataSourcePluginOptionsEditorProps } from '@grafana/data';
 import { AxiomDataSourceOptions, MySecureJsonData } from '../types';
 
@@ -7,6 +7,10 @@ interface Props extends DataSourcePluginOptionsEditorProps<AxiomDataSourceOption
 
 export function ConfigEditor(props: Props) {
   const { onOptionsChange, options } = props;
+  const [shouldShowOrgId, setShowOrgId] = useState(
+    !!options.jsonData.orgID && options.secureJsonData?.accessToken.startsWith('xapt-')
+  );
+
   const onHostChange = (event: ChangeEvent<HTMLInputElement>) => {
     const jsonData = {
       ...options.jsonData,
@@ -15,16 +19,14 @@ export function ConfigEditor(props: Props) {
     onOptionsChange({ ...options, jsonData });
   };
 
-  const onOrgIDChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const jsonData = {
-      ...options.jsonData,
-      orgID: event.target.value,
-    };
-    onOptionsChange({ ...options, jsonData });
-  };
-
   // Secure field (only sent to the backend)
   const onAccessTokenChange = (event: ChangeEvent<HTMLInputElement>) => {
+    if (event.target.value.startsWith('xapt-')) {
+      setShowOrgId(true);
+    } else {
+      setShowOrgId(false);
+    }
+
     onOptionsChange({
       ...options,
       secureJsonData: {
@@ -47,30 +49,52 @@ export function ConfigEditor(props: Props) {
     });
   };
 
+  const onOrgIDChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const jsonData = {
+      ...options.jsonData,
+      orgID: event.target.value,
+    };
+    onOptionsChange({ ...options, jsonData });
+  };
+
   const { secureJsonFields } = options;
   const jsonData = (options.jsonData || {}) as AxiomDataSourceOptions;
   const secureJsonData = (options.secureJsonData || {}) as MySecureJsonData;
 
   return (
     <div className="gf-form-group">
-      <Label description={<span>Create a Personal Token from your Axiom account settings.</span>}>
+      <Label description={<span>Create an API Token from your Axiom account settings.</span>}>
         <h5>Authentication</h5>
       </Label>
-      <InlineField label="Personal Token" labelWidth={17}>
+      <InlineField label="API Token" labelWidth={17}>
         <SecretInput
           isConfigured={(secureJsonFields && secureJsonFields.accessToken) as boolean}
           value={secureJsonData.accessToken || ''}
-          placeholder="xapt-***********"
+          placeholder="xaat-***********"
           width={40}
           onReset={onResetAccessToken}
           onChange={onAccessTokenChange}
         />
       </InlineField>
       <br />
-      <InlineField label="Org ID" labelWidth={17}>
-        <Input value={jsonData.orgID || ''} placeholder="" width={40} onChange={onOrgIDChange} />
-      </InlineField>
-      <br />
+      {/* Only show orgId for users who have already set it. Promote advanced tokens instead */}
+      {shouldShowOrgId && (
+        <InlineField label="Org ID" labelWidth={17}>
+          <Input value={jsonData.orgID || ''} placeholder="" width={40} onChange={onOrgIDChange} />
+        </InlineField>
+      )}
+      {/* If orgId is set, show a deprecation message */}
+      {shouldShowOrgId && (
+        <div>
+          <Alert
+            title="Personal tokens are deprecated and will be removed in the next release. Please switch to advanced API tokens."
+            about="Token"
+            severity="warning"
+            buttonContent="Learn more"
+            topSpacing={4}
+          />
+        </div>
+      )}
       <div>
         <Label description="The Axiom host to use.">
           <h6>Axiom Host</h6>